Put uart_bind() into uart_common.c
