+++ client/Makefile
@@ -713,6 +713,7 @@ SRCS =  mifare/aiddesfire.c \
 		proxmark3.c \
 		scandir.c \
 		uart/ringbuffer.c \
+		uart/uart_common.c \
 		uart/uart_posix.c \
 		uart/uart_win32.c \
 		scripting.c \
